Seperated the functions into their own files, updated gitignore
