+++ vendor_backups/vyos.py
@@ -1,0 +1,32 @@
+from netmiko import ConnectHandler
+from datetime import datetime
+
+# Current time and formats it to the North American time of Month, Day, and Year.
+now = datetime.now()
+dt_string = now.strftime("%m-%d-%Y_%H-%M")
+
+# Gives us the information we need to connect to VyOS devices.
+def get_saved_config_vyos(host, username, password):
+    vyos = {
+        'device_type': 'vyos',
+        'host': host,
+        'username': username,
+        'password': password,
+    }
+    # Creates the connection to the device.
+    net_connect = ConnectHandler(**vyos)
+    net_connect.enable()
+    # Gets the running configuration.
+    output = net_connect.send_command("show conf comm")
+    # Gets and splits the hostname for the output file name.
+    hostname = net_connect.find_prompt().replace('#','').replace('>','')
+    if not hostname:
+        hostname = net_connect.send_command("sh conf | grep host-name | awk {'print $2'}")
+        hostname = hostname.split()
+        hostname = hostname[0]
+    # Creates the file name, which is the hostname, and the date and time.
+    fileName = hostname + "_" + dt_string
+    # Creates the text file in the backup-config folder with the special name, and writes to it.
+    backupFile = open("backup-config/" + fileName + ".txt", "w+")
+    backupFile.write(output)
+    print("Outputted to " + fileName + ".txt")
